requirements.txt
click>=8.0.0
pyyaml>=6.0
colorama>=0.4.4
tabulate>=0.9.0
python-dateutil>=2.8.0
rich>=13.0.0
textual>=0.44.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=5.0.0
mypy>=1.0.0